CoordinateBisection

Documentation for CoordinateBisection.

CoordinateBisection.RCB

CoordinateBisection.domain_weights

CoordinateBisection.RCB — Type

struct RCB{N, T<:Real}

Split a given set of points with optional weights into N balanced axis-aligned domains. Balanced means:

If the weights are nothing, the number of points in each domain is equalised.

If the weights are a Vector for each point, the sum of weights on each domain is equalised.

If the weights are a Matrix for each dimension of each 3D point, the sum of coordinate-wise weights is equalised.

The resulting RCB tree stores each domain's neighbours in the x, y and z members, each domain's span in the xspan, yspan and zspan members, and the spatial limits of the points used to construct it in the xlimit, ylimit and zlimit members.

The input points should have shape (3, NumPoints). When computing the neighbours, an optional "skin" distance may be specified, such that if any domains are less than "skin" apart (even if they are not physically touching), they are considered neighbours.

Once the initial RCB tree is constructed, it can be iteratively improved in optimize iterations. This is a mathematically terrible problem, so no absolute guarantees can be offered at the moment as to whether they will always improve the splitting; however, in all our tests - even with skewed points distributions - it behaves well.

Fields

x::NTuple{N, Array{Tuple{Int64, T}, 1}} where {N, T<:Real}: Each domain's neighbours in the X/Y/Z dimension, as Tuple{DomainIndex, SplitCoordinate}

y::NTuple{N, Array{Tuple{Int64, T}, 1}} where {N, T<:Real}

z::NTuple{N, Array{Tuple{Int64, T}, 1}} where {N, T<:Real}

xspan::NTuple{N, Tuple{T, T}} where {N, T<:Real}: Each domain's span in the X/Y/Z dimension, as Tuple{MinCoordinate, MaxCoordinate}

yspan::NTuple{N, Tuple{T, T}} where {N, T<:Real}

zspan::NTuple{N, Tuple{T, T}} where {N, T<:Real}

xlimit::Tuple{T, T} where T<:Real: Limits of points over which the RCB was constructed, as Tuple{MinCoordinate, MaxCoordinate}

ylimit::Tuple{T, T} where T<:Real

zlimit::Tuple{T, T} where T<:Real

Methods

RCB{N, T}(points, weights=nothing, skin=T(0.), optimize=2N) where {N, T <: Real}
ndims(::RCB{N, T}) where {N, T} = N
eltype(::RCB{N, T}) where {N, T} = T

Examples

Split 10 points into 3 domains:

using CoordinateBisection
using Random

Random.seed!(123)
points = rand(3, 10)

rcb = RCB{3, Float64}(points)

You can plot results if Makie is loaded:

using GLMakie

# Possible rcbplot arguments:
rcbplot(rcb)
rcbplot(rcb, points)
rcbplot(rcb, points, weights)

Interactive plots showing optimisation iterations can be shown with GLMakie:

using GLMakie

interactive_rcbplot(points, weights)

CoordinateBisection.domain_weights — Method

domain_weights(
    rcb::RCB{N, T},
    points,
    weights=nothing,
) where {N, T} -> MVector{N, Tuple{T, T, T}}

Compute the total points' weight in each domain of rcb.